working commit
updated glew and glfw
added glew dynamically (not statically anymore)
checking if file .shader file is available and print message if it doesn't exist.
(glfw actually needed to be updated, because it was a new version for vs2022)
brought MapNamedBuffer.cpp finally to work!! yay!
made summary of learned stuff in readme.txt